validaciones

+++ http/wsconfig/http.ps1
@@ -76,17 +76,25 @@ while($true){
                 "1"{
                     try {
                         $puerto = Read-Host "Ingresa el puerto donde se realizara la instalacion"
-                        Stop-Process -Name nginx -ErrorAction SilentlyContinue
-                        echo "Instalando version LTS $versionLTSNginx"
-                        Invoke-WebRequest -UseBasicParsing "https://nginx.org/download/nginx-$versionLTSNginx.zip" -Outfile "C:\descargas\nginx-$versionLTSNginx.zip"
-                        Expand-Archive C:\descargas\nginx-$versionLTSNginx.zip C:\descargas -Force | Out-Null
-                        cd C:\descargas\nginx-$versionLTSNginx
-                        Start-Process nginx.exe
-                        Get-Process | Where-Object { $_.ProcessName -like "*nginx*" }
-                        cd ..
-                        (Get-Content C:\descargas\nginx-$versionLTSNginx\conf\nginx.conf) -replace "listen       [0-9]{1,5}", "listen       $puerto" | Set-Content C:\descargas\nginx-$versionLTSNginx\conf\nginx.conf
-                        Select-String -Path "C:\descargas\nginx-$versionLTSNginx\conf\nginx.conf" -Pattern "listen       [0-9]{1,5}"
-                        echo "Se instalo la version LTS $versionLTSNginx de Nginx"
+                        if(-not(Es-PuertoValido -puerto $puerto)){
+                            echo "Ingresa un puerto dentro del rango (1024-65535)"
+                        }
+                        elseif(-not(Es-Numerico -string $puerto)){
+                            echo "Ingresa un valor numerico entero"
+                        }
+                        else{
+                            Stop-Process -Name nginx -ErrorAction SilentlyContinue
+                            echo "Instalando version LTS $versionLTSNginx"
+                            Invoke-WebRequest -UseBasicParsing "https://nginx.org/download/nginx-$versionLTSNginx.zip" -Outfile "C:\descargas\nginx-$versionLTSNginx.zip"
+                            Expand-Archive C:\descargas\nginx-$versionLTSNginx.zip C:\descargas -Force
+                            cd C:\descargas\nginx-$versionLTSNginx
+                            Start-Process nginx.exe
+                            Get-Process | Where-Object { $_.ProcessName -like "*nginx*" }
+                            cd ..
+                            (Get-Content C:\descargas\nginx-$versionLTSNginx\conf\nginx.conf) -replace "listen       [0-9]{1,5}", "listen       $puerto" | Set-Content C:\descargas\nginx-$versionLTSNginx\conf\nginx.conf
+                            Select-String -Path "C:\descargas\nginx-$versionLTSNginx\conf\nginx.conf" -Pattern "listen       [0-9]{1,5}"
+                            echo "Se instalo la version LTS $versionLTSNginx de Nginx"
+                        }
                     }
                     catch {
                         Echo $Error[0].ToString()
@@ -94,18 +102,26 @@ while($true){
                 }
                 "2"{
                     try {
-                        Stop-Process -Name nginx -ErrorAction SilentlyContinue
                         $puerto = Read-Host "Ingresa el puerto donde se realizara la instalacion"
-                        echo "Instalando version de desarrollo $versionDevNginx"
-                        Invoke-WebRequest -UseBasicParsing "https://nginx.org/download/nginx-$versionDevNginx.zip" -Outfile "C:\descargas\nginx-$versionDevNginx.zip"
-                        Expand-Archive C:\descargas\nginx-$versionDevNginx.zip C:\descargas -Force | Out-Null
-                        cd C:\descargas\nginx-$versionDevNginx
-                        Start-Process nginx.exe
-                        Get-Process | Where-Object { $_.ProcessName -like "*nginx*" }
-                        cd ..
-                        (Get-Content C:\descargas\nginx-$versionDevNginx\conf\nginx.conf) -replace "listen       [0-9]{1,5}", "listen       $puerto" | Set-Content C:\descargas\nginx-$versionDevNginx\conf\nginx.conf
-                        Select-String -Path "C:\descargas\nginx-$versionDevNginx\conf\nginx.conf" -Pattern "listen       [0-9]{1,5}"
-                        echo "Se instalo la Version de desarrollo $versionDevNginx de Nginx"
+                        if(-not(Es-PuertoValido -puerto $puerto)){
+                            echo "Ingresa un puerto dentro del rango (1024-65535)"
+                        }
+                        elseif(-not(Es-Numerico -string $puerto)){
+                            echo "Ingresa un valor numerico entero"
+                        }
+                        else{
+                            Stop-Process -Name nginx -ErrorAction SilentlyContinue
+                            echo "Instalando version de desarrollo $versionDevNginx"
+                            Invoke-WebRequest -UseBasicParsing "https://nginx.org/download/nginx-$versionDevNginx.zip" -Outfile "C:\descargas\nginx-$versionDevNginx.zip"
+                            Expand-Archive C:\descargas\nginx-$versionDevNginx.zip C:\descargas -Force
+                            cd C:\descargas\nginx-$versionDevNginx
+                            Start-Process nginx.exe
+                            Get-Process | Where-Object { $_.ProcessName -like "*nginx*" }
+                            cd ..
+                            (Get-Content C:\descargas\nginx-$versionDevNginx\conf\nginx.conf) -replace "listen       [0-9]{1,5}", "listen       $puerto" | Set-Content C:\descargas\nginx-$versionDevNginx\conf\nginx.conf
+                            Select-String -Path "C:\descargas\nginx-$versionDevNginx\conf\nginx.conf" -Pattern "listen       [0-9]{1,5}"
+                            echo "Se instalo la Version de desarrollo $versionDevNginx de Nginx"
+                        }
                     }
                     catch {
                         echo $Error[0].ToString()
